django_drf_nuxt_services_aggregator/backend/apps/company/serializers/brand.py
from rest_framework import serializers

from apps.company.models.brand import Brand


class BrandSerializer(serializers.ModelSerializer):
    class Meta:
        model = Brand
        fields = '__all__'
