django_drf_nuxt_services_aggregator/backend/requirements/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.6
django-appconf==1.0.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-imagekit==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
kombu==5.2.4
packaging==21.3
pilkit==2.0
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
